Move the CommandSockets to its own module
