+++ silex_client/action/abstract_socket.py
@@ -4,13 +4,16 @@
 
 Class definition of AbstractSocketBuffer
 """
+from __future__ import annotations
 
 from abc import ABC, abstractmethod
 from dataclasses import dataclass, field
-from typing import Any, Union
+from typing import Any, Union, TYPE_CHECKING
 
 from silex_client.utils.socket_types import SocketTypeMeta
 
+if TYPE_CHECKING:
+    from silex_client.action.action_query import ActionQuery
 
 # Alias the metaclass type, to avoid clash with the type attribute
 TypeAlias = type
@@ -33,7 +36,7 @@ class AbstractSocketBuffer(ABC):
     hide: bool = field(compare=False, repr=False, default=False)
 
     @abstractmethod
-    def eval(self, action_query) -> Any:
+    def eval(self, action_query: ActionQuery) -> Any:
         """
         The output of a socket buffer is the result of the input casted to the desired type
         """
