+++ silex_client/action/socket_buffer.py
@@ -4,7 +4,6 @@
 
 Class definition of SocketBuffer
 """
-
 from __future__ import annotations
 
 from dataclasses import dataclass, field
